refactor(web): lift the composer's uploads and readouts out of the input
The picker limits were the store's own defaults; the composer stops
restating them, and the command-limit assertion moves onto the frame the
store actually sends.
